fix: ensure min size properties do not affect stack mode
